+++ utils/travis-run-install.sh
@@ -1,7 +1,8 @@
 set -e
 set -x
 
-export PATH="~/miniconda/bin:$PATH"
+export INSTALL_PREFIX=~/miniconda
+export PATH=$INSTALL_PREFIX/bin:$PATH
 
 
 osx_setup() {
@@ -9,33 +10,6 @@ osx_setup() {
 
     brew outdated openssl || brew upgrade openssl
     brew install zsh
-
-#    # install pyenv
-#    git clone https://github.com/yyuu/pyenv.git ~/.pyenv
-#    PYENV_ROOT="$HOME/.pyenv"
-#    PATH="$PYENV_ROOT/bin:$PATH"
-#    eval "$(pyenv init -)"
-#
-#    case "$PYTHON_VERSION" in
-#        '2.7')
-#            curl -O https://bootstrap.pypa.io/get-pip.py
-#            python get-pip.py --user
-#            ;;
-#        '3.4')
-#            pyenv install 3.4.5
-#            pyenv global 3.4.5
-#            ;;
-#        '3.5')
-#            pyenv install 3.5.2
-#            pyenv global 3.5.2
-#            ;;
-#        '3.6')
-#            pyenv install 3.6.0
-#            pyenv global 3.6.0
-#            ;;
-#    esac
-#    pyenv rehash
-#    export PYTHON_EXE="$(pyenv which python)"
 
     rvm get head
 }
@@ -45,21 +19,23 @@ install_python() {
     # strategy is to use Miniconda to install python, but then remove all vestiges of conda
    curl -sSL $MINICONDA_URL -o ~/miniconda.sh
    chmod +x ~/miniconda.sh
-   ~/miniconda.sh -bfp ~/miniconda
+   mkdir -p $INSTALL_PREFIX
+   ~/miniconda.sh -bfp $INSTALL_PREFIX
    hash -r
-   conda install -y -q python=$PYTHON_VERSION
-   local site_packages=$(~/miniconda/bin/python -c "from distutils.sysconfig import get_python_lib as g; print(g())")
-   rm -rf ~/miniconda/bin/activate \
-       ~/miniconda/bin/conda \
-       ~/miniconda/bin/deactivate \
-       ~/miniconda/conda-meta/conda-*.json \
-       ~/miniconda/conda-meta/requests-*.json \
-       ~/miniconda/conda-meta/pyopenssl-*.json \
-       ~/miniconda/conda-meta/cryptography-*.json \
-       ~/miniconda/conda-meta/idna-*.json \
-       ~/miniconda/conda-meta/ruamel-*.json \
-       ~/miniconda/conda-meta/pycrypto-*.json \
-       ~/miniconda/conda-meta/pycosat-*.json \
+   $INSTALL_PREFIX/bin/conda install -y -q python=$PYTHON_VERSION
+   local site_packages=$($INSTALL_PREFIX/bin/python -c "from distutils.sysconfig import get_python_lib as g; print(g())")
+   rm -rf $INSTALL_PREFIX/bin/activate \
+       $INSTALL_PREFIX/bin/conda \
+       $INSTALL_PREFIX/bin/conda-env \
+       $INSTALL_PREFIX/bin/deactivate \
+       $INSTALL_PREFIX/conda-meta/conda-*.json \
+       $INSTALL_PREFIX/conda-meta/requests-*.json \
+       $INSTALL_PREFIX/conda-meta/pyopenssl-*.json \
+       $INSTALL_PREFIX/conda-meta/cryptography-*.json \
+       $INSTALL_PREFIX/conda-meta/idna-*.json \
+       $INSTALL_PREFIX/conda-meta/ruamel-*.json \
+       $INSTALL_PREFIX/conda-meta/pycrypto-*.json \
+       $INSTALL_PREFIX/conda-meta/pycosat-*.json \
        $site_packages/conda* \
        $site_packages/requests* \
        $site_packages/pyopenssl* \
@@ -69,50 +45,44 @@ install_python() {
        $site_packages/pycrypto* \
        $site_packages/pycosat*
    hash -r
-   which -a python
-   ~/miniconda/bin/python --version
-   ~/miniconda/bin/pip --version
+   $INSTALL_PREFIX/bin/python --version
+   $INSTALL_PREFIX/bin/pip --version
 
 }
 
 miniconda_install() {
     curl -L https://repo.continuum.io/miniconda/Miniconda3-4.0.5-Linux-x86_64.sh -o ~/miniconda.sh
-    bash ~/miniconda.sh -bfp ~/miniconda
-    export PATH=~/miniconda/bin:$PATH
+    bash ~/miniconda.sh -bfp $INSTALL_PREFIX
     hash -r
-    which -a conda
-    conda install -y -q pip conda 'python>=3.6'
-    conda info
-    which -a pip
-    which -a python
-    conda config --set auto_update_conda false
+    $INSTALL_PREFIX/bin/conda install -y -q pip conda 'python>=3.6'
+    $INSTALL_PREFIX/bin/conda info
+    $INSTALL_PREFIX/bin/conda config --set auto_update_conda false
 }
 
 
 conda_build_install() {
     # install conda
-    rm -rf $(~/miniconda/bin/python -c "import site; print(site.getsitepackages()[0])")/conda
-    ~/miniconda/bin/python utils/setup-testing.py install
-    hash -r
-    conda info
+    $INSTALL_PREFIX/bin/pip install -r utils/requirements-test.txt
+    $INSTALL_PREFIX/bin/python utils/setup-testing.py install
+    $INSTALL_PREFIX/bin/conda info
 
     # install conda-build test dependencies
-    conda install -y -q pytest pytest-cov pytest-timeout mock
-    conda install -y -q -c conda-forge perl pytest-xdist
-    conda install -y -q anaconda-client numpy
+    $INSTALL_PREFIX/bin/conda install -y -q pytest pytest-cov pytest-timeout mock
+    $INSTALL_PREFIX/bin/conda install -y -q -c conda-forge perl pytest-xdist
+    $INSTALL_PREFIX/bin/conda install -y -q anaconda-client numpy
 
-    ~/miniconda/bin/python -m pip install pytest-catchlog pytest-mock
+    $INSTALL_PREFIX/bin/pip install pytest-catchlog pytest-mock
 
-    conda config --set add_pip_as_python_dependency true
+    $INSTALL_PREFIX/bin/conda config --set add_pip_as_python_dependency true
 
     # install conda-build runtime dependencies
-    conda install -y -q filelock jinja2 patchelf conda-verify setuptools contextlib2 pkginfo
+    $INSTALL_PREFIX/bin/conda install -y -q filelock jinja2 patchelf conda-verify setuptools contextlib2 pkginfo
 
     # install conda-build
     git clone -b $CONDA_BUILD --single-branch --depth 1000 https://github.com/conda/conda-build.git
-    rm -rf $(~/miniconda/bin/python -c "import site; print(site.getsitepackages()[0])")/conda_build
+    rm -rf $($INSTALL_PREFIX/bin/python -c "import site; print(site.getsitepackages()[0])")/conda_build
     pushd conda-build
-    ~/miniconda/bin/pip install .
+    $INSTALL_PREFIX/bin/pip install .
     hash -r
     conda info
     popd
@@ -135,12 +105,17 @@ esac
 
 
 if [[ $FLAKE8 == true ]]; then
-    install_python
-    ~/miniconda/bin/pip install flake8
+    pip install flake8
+elif [[ $TRAVIS_SUDO == true ]]; then
+    sudo -E -u root bash -c "source utils/functions.sh && install_conda_dev /usr/local"
+    export INSTALL_PREFIX="/usr/local"
+    export PATH=$INSTALL_PREFIX/bin:$PATH
+    sudo -E -u root chown -R root:root ./conda
+    ls -al ./conda
 elif [[ -n $CONDA_BUILD ]]; then
-    miniconda_install
+    install_python
     conda_build_install
 else
     install_python
-    ~/miniconda/bin/pip install -r utils/requirements-test.txt
+    $INSTALL_PREFIX/bin/pip install -r utils/requirements-test.txt
 fi